Changed tests and changed type of primary key from int to long

+++ src/main/java/com/kosmoastronauta/demo/repository/BookRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends CrudRepository<Book, Integer>, BookRepositoryCustom
+public interface BookRepository extends CrudRepository<Book, Long>, BookRepositoryCustom
 {
 }
